depth only rendering
git-svn-id: https://g3d.svn.sourceforge.net/svnroot/g3d@749 ef0428c2-0c16-0410-9f3f-d710ca1f8061

+++ G3D9/GLG3D.lib/source/VertexRange.cpp
@@ -193,7 +193,7 @@ void VertexRange::set(int index, const void* value, GLenum glformat, int eltSize
     debugAssertM(glformat == m_underlyingRepresentation, 
         "Value argument to VertexRange::set must match the intialization type.");
 
-    debugAssertM(eltSize == m_elementSize, 
+    debugAssertM((size_t)eltSize == m_elementSize, 
         "Value argument to VertexRange::set must match the intialization type's memory footprint.");
 
     uploadToCard(value, index * eltSize, eltSize);
